adding Managers with functionality

+++ Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib/receiverBoard.cpp
@@ -4,7 +4,7 @@
 #include "receiverBoard.h"
 
 
-   int getBoard_ID() {
+   unsigned long int getBoard_ID() {
      return ESP.getEfuseMac();
    }
