+++ Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib/receiverBoard.h
@@ -29,6 +29,6 @@
 
 
 
-   int getBoard_ID();
+   unsigned long int getBoard_ID();
 
 #endif
